Improve tests; fix bugs found by improved tests
I've added some objdump disassembly of an esp32 ELF I had laying around;
this tests (I assume...) the ESP32-specific instructions added earlier.

Also bzipped the test data, to save a bit on disk space. It was only a
couple megs, but now it's less than half a meg :D

TIL that Python will let you open a bzipped file as if it's a normal
file with bz2.open, so I did that.