Fix disassembly of ROTW instruction
ROTW has an imm4 value that we weren't properly handling. We ended up
returning an InstructionTextToken with value None, which caused Binary
Ninja to crash.

imm4's don't have a standardized location, so I've introduced a
`rotw_simm4` function to get the immediate to ROTW if we need it in the
disassembler and lifter.

Additionally, tests have been added to ensure this issue does not
reoccur. I assembled the test case by hand, so if there are any issues
that's where they came from.